+++ pinot-spi/src/main/java/org/apache/pinot/spi/annotations/segment/lifecycle/SegmentDeletionListener.java
@@ -1,5 +1,3 @@
-
-
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
  * or more contributor license agreements.  See the NOTICE file
